// projects/cashier/src/lib/models/response/cashierdetails.response.model.ts
export interface CasiherDataResponse {
  vip: number;
  ib: number;
  vipIb: number;
  tbChips: number;
  viptb: number;
  freeroll: number;
  booster: number;
  crowns: number;
  depositAmt: number;
  withdrawlAmt: number;
  remainingTime: number;
  name: string;
  userName: string;
  holdAmt: number;
  realChips: number;
  total: number;
  totalbonus: number;
}